SpectroPipeR: MVA module

Function for conducting multivariate analysis, which represents the third step in the pipeline and builds upon the normalization and quantification module from step 2.

Usage

MVA_module(
  SpectroPipeR_data_quant = NULL,
  HCPC_analysis = FALSE,
  costum_colors = NULL
)

Arguments

SpectroPipeR_data_quant

it is the SpectroPipeR_data_quant list object from norm_quant_module() object e.g. data_input_norm_quant see example below

HCPC_analysis

boolean; should a HCPC be performed or not

costum_colors

if you would like to use your own colors please provide a named color vector (e.g. c(condition1 = "black", condition2 = "grey")); names should have the same naming and length like the conditions set in Spectronaut

Value

saves MVA analysis results in output folder

SpectroPipeR_data_MVA list object with the MVA analysis data in addition to the automatically saved tables and plots For the description of the generated figures and tables please read the manual & vignettes

list element 	 description

PCA_peptide_intensity	PCA list element: PCA list element of peptide int.

PCA_protein_intensity	PCA list element: PCA list element of protein int.

UMAP_protein_intensity	umap element: UMAP element of protein int.

peptide_intensity_correlation	matrix: Spearman correlation scores of peptide int.

protein_intensity_correlation	matrix: Spearman correlation scores of protein int.

Examples

# \donttest{
#load library
library(SpectroPipeR)

# use default parameters list
params <- list(output_folder = "../SpectroPipeR_test_folder")

# example input file
example_file_path <- system.file("extdata",
                                "SN_test_HYE_mix_file.tsv",
                                package="SpectroPipeR")

# step 1: load Spectronaut data module
SpectroPipeR_data <- read_spectronaut_module(file = example_file_path,
                                            parameter = params,
                                            print.plot = FALSE)

# step 2: normalize & quantification module
SpectroPipeR_data_quant <- norm_quant_module(SpectroPipeR_data = SpectroPipeR_data)

# step 3: MVA module
SpectroPipeR_MVA <- MVA_module(SpectroPipeR_data_quant = SpectroPipeR_data_quant,
          HCPC_analysis = FALSE)
# }