SpectroPipeR: complete Spectronaut analysis with one function

Function for performing the whole SpectroPipeR analysis workflow.

Usage

SpectroPipeR(
  file = "",
  parameter = list(),
  max_chars_file_name_capping = 25,
  ID_condition_filtering = FALSE,
  ID_condition_filtering_percent = 0.5,
  batch_adjusting = FALSE,
  sample__batch_meta_data_file = NULL,
  batch_adjusting_column = "",
  number_of_cores_adjusting = parallel::detectCores() - 2,
  covariate_adjusting_formula = "",
  covariate_adjusting_meta_data_file = "",
  skipping_MaxLFQ_median_norm = FALSE,
  HCPC_analysis = FALSE,
  costum_colors = NULL,
  condition_comparisons = NULL,
  number_of_cores_statistics = 2,
  build_HTML_report = TRUE
)

Arguments

file

location (path) of Spectronaut output report; you should use the Spectronaut_export_scheme() function for getting a SpectroPipeR report scheme encompassing all mandatory columns

parameter

mandatory parameter list element

table of list elements:

parameter 	 description

output_folder	character - output folder path (abs.)

ion_q_value_cutoff	numeric - Q-value used in Spectronaut analysis: Biognosys

default is 0.01 = 1% error rate

id_drop_cutoff	numeric - value between 0-1 (1 = 100%); xx percent lower

than median of ion ID rate => outlier

normalization_method	character - "median" or Spectronaut - auto-detection is per

default ON, meaning if normalization was performed in Spectronaut

this will be detected and preferred over parameter setting here;

median normalization is the fallback option

normalization_factor_cutoff_outlier	numeric - median off from global median

(4 means abs. 4fold off)

filter_oxidized_peptides	logical - if oxidized peptides should be removed from peptide

quantification

protein_intensity_estimation	character - Hi3 = Hi3 protein intensity estimation,

MaxLFQ = MaxLFQ protein intensity estimation

stat_test	character - choose statistical test: "rots" = reproducibility

optimized test statistics, "modt" = moderate t-test (lmfit, eBayes),

"t" = t-test

type_slr	character - choose ratio aggregation method: "median" or "tukey"

is used when calculating protein values

fold_change	numeric - fold-change used as cutoff e.g. 1.5

p_value_cutoff	numeric - p-value used as cutoff e.g. 0.05

paired	logical - Should paired statistics be applied?

example parameters list:

params <- list(output_folder = "../Spectronaut_example",

ion_q_value_cutoff = 0.001,

id_drop_cutoff = 0.3,

normalization_method = "median",

normalization_factor_cutoff_outlier = 4,

filter_oxidized_peptides = T,

protein_intensity_estimation = "MaxLFQ",

stat_test = "modt",

type_slr = "median",

fold_change = 1.5,

p_value_cutoff = 0.05,

paired = FALSE

)

max_chars_file_name_capping

integer, (default = 25) number of max characters used for raw file name presentation; must be adjusted if function

ID_condition_filtering

TRUE or FALSE if a condition-wise filtering should be performed

ID_condition_filtering_percent

(numerical value ranging from 0 - 1, default = 0.5) define the proportion for the condition-wise ID filtering

batch_adjusting

logical - if batch adjusting with ComBat (sva package) should be performed; default = FALSE

sample__batch_meta_data_file

character - sample batch file; tab-delimited txt-file, containing "R.FileName" column e.g. sample__batch_meta_data_file = "Sample_MetaData_Batches.txt"

example table for batch meta data:

R.FileName 	 digest_batch

20230403_TIMSTOF_1_S1-B11_1_6690	1

20230403_TIMSTOF_2_S1-G11_1_6695	1

20230403_TIMSTOF_3_S1-E7_1_6661	1

20230403_TIMSTOF_4_S1-A9_1_6673	2

20230403_TIMSTOF_7_S1-D8_1_6668	2

20230403_TIMSTOF_9_S1-D3_1_6627	2

A good starting point for the generation of the table is the '*_ConditionSetup.tsv' in your Spectronaut Pipeline Report export folder

batch_adjusting_column

character - column name in sample__batch_meta_data_file, which should be used for assigning the samples to batches

number_of_cores_adjusting

numeric - number of processor cores used for batch or covariate adjustment

covariate_adjusting_formula

character - provide a formula passed to lm() for covariate adjustment e.g. "log10_peptide_intensity ~ log10(CRP)+log10(age)+as.factor(sex)"; you may also use ns() function e.g. "log10_peptide_intensity ~ ns(age, df=3)"

covariate_adjusting_meta_data_file

covariate meta csv file, containing "R.FileName; age; sex;..."; you may find a start file in the 02_ID_rate folder > file_list.csv column e.g. covariate_adjusting_meta_data_file = "covariate_MetaData_file.csv" example table for covariate meta data:

R.FileName 	 R.Condition 	 sex 	 CRP

20230403_TIMSTOF_1_S1-B11_1_6690	heathy	1	3.8

20230403_TIMSTOF_2_S1-G11_1_6695	heathy	2	5.1

20230403_TIMSTOF_3_S1-E7_1_6661	heathy	1	1.2

20230403_TIMSTOF_4_S1-A9_1_6673	cancer	1	50.2

20230403_TIMSTOF_7_S1-D8_1_6668	cancer	2	30.8

20230403_TIMSTOF_9_S1-D3_1_6627	cancer	2	64.1

skipping_MaxLFQ_median_norm

logical - if median normalization after MaxLFQ calculation should be skipped; default = FALSE; applied only if MaxLFQ protein estimation is selected

HCPC_analysis

boolean; should a HCPC be performed or not

costum_colors

if you would like to use your own colors for condition coloring please provide a named color vector (e.g. c(condition1 = "black", condition2 = "grey")); names should have the same naming and length like the conditions set in Spectronaut

condition_comparisons

condition comparisons for pairwise- comparison; e.g. condition_comparisons <- cbind(c("condition1","control"),c("condition3","control") )

number_of_cores_statistics

number of processor cores to be used for the calculations default = 2;

parallel::detectCores()-2 for faster processing (will detect the number of cores in the system and use nearly all cores)

build_HTML_report

boolean; if a HTML report of the analysis should be generated or not

Value

SpectroPipeR list object containing tables and plots of the analysis in addition to the automatically saved tables and plots. For the description of the generated figures and tables please read the manual & vignettes

The SpectroPipeR list element contains:

SpectroPipeR_data

SpectroPipeR_data_quant

SpectroPipeR_data_MVA

SpectroPipeR_data_stats

SpectroPipeR_data:

list element 	 description

spectronaut_output	tibble: Spectronaut report tibble provided for the analysis

SDRF_file	tibble: intermediate SDRF table of the analysis

summary_distinct	tibble: distinct ion, modified peptide, stripped peptides and

protein group count per file filtered by provided Q-value

raw_file_names	tibble: R.FileNames capped and uncapped version together with

R.Condition and R.Replicate

ion_id_median	numerical value: median of ion intensity

ion_id_cutoff	numerical value: ion ID count threshold to classify sample as outlier

PG_2_peptides_ID_raw	tibble: with protein groups with at least 2 peptides with peptide

and replicate count

summary_distinct_outlier	tibble: if outlier are detected they are listed in this tibble

ID_rate_plot	ggplot2 plot: ID rate plot

ID_rate_plot_filter	ggplot2 plot: ion ID rate plot with ion ID cutoff line

sample_length	numberical value: number of samples in the provided Spectronaut report

parameter	list: parameters provided by the user

time_stamp_log_file	string: time stamp of the log file (format: %Y_%m_%d__%H_%M)

log_file_name	string: analysis log file name

SpectroPipeR_data_quant:

list element 	 description

data_input_normalized	tibble: Spectronaut report tibble provided for the analysis

MedianNormalizationFactor	tibble: ion normalization factor table

MedianNormalizationFactor_outlier	tibble: table containing detected norm. outliers on ion level

NormFactor_plot	ggplot2 plot: norm. factor plots

iBAQ_intensities	tibble: table containing the iBAQ int.

iBAQ_intensities_summary	tibble: table containing the per condition summarized iBAQ int.

protein_data	tibble: protein intensity table (e.g. Hi3 or MaxLFQ int.)

PG_2_peptides_ID_raw	tibble: with protein groups with at least 2 peptides with peptide

and replicate count

protein_data_normalization_factor	tibble: normalization factor table for protein int. data

peptide_intensity_filtered_2pep_hi3	tibble: if Hi3 protein int. was selected a table containing

the peptides and intensities used for Hi3 protein intensity calculation

peptide_intensity	tibble: peptides intensity table based on norm. ion intensity

parameter	list: parameters provided by the user updated with the

norm_quant_module() parameters

CV_cumulative_frequency	tibble: cumulative frequency table on peptide

and protein intensity level

sample_length	numberical value: number of samples in the provided Spectronaut report

SpectroPipeR_data_MVA:

list element 	 description

PCA_peptide_intensity	PCA list element: PCA list element of peptide int.

PCA_protein_intensity	PCA list element: PCA list element of protein int.

UMAP_protein_intensity	umap element: UMAP element of protein int.

peptide_intensity_correlation	matrix: Spearman correlation scores of peptide int.

protein_intensity_correlation	matrix: Spearman correlation scores of protein int.

SpectroPipeR_data_stats:

list element 	 description

stat_results	tibble: statistical analysis results table

stat_column_description	tibble: statistical analysis results table column description

stats_results_iBAQ_quantiles	tibble: statistical analysis results table containing

the iBAQ quantilies (Q1-Q10) of the protein per group for a better

ratio judgement

stat_results_filtered	tibble: filtered (user defined FC and adj. p-value) statistical

analysis results table

Details

batch adjustment

Batch effects refer to systematic differences between batches (groups) of samples in high-throughput experiments. These differences can arise due to various factors, such as batch variations in sample preparation, handling, processing procedures and measurement orders. Batch effects can obscure the true biological signal and lead to incorrect conclusions if not properly accounted for. In the SpectroPipeR pipeline, the ComBat tool was employed to adjust for batch effects in the datasets where the batch covariate was known. ComBat utilizes the methodology described in Johnson et al. 2007. It uses an empirical Bayes (EB) framework for adjusting data for batch effects that is robust to outliers in small sample sizes and performs comparable to existing methods for large samples. Johnson et al. 2007: This method incorporates systematic batch biases common across genes in making adjustments, assuming that phenomena resulting in batch effects often affect many genes in similar ways (i.e. increased expression, higher variability, etc). Specifically, the the L/S model parameters are estimated that represent the batch effects by pooling information across peptides in each batch to shrink the batch effect parameter estimates toward the overall mean of the batch effect estimates (across genes). These EB estimates are then used to adjust the data for batch effects, providing more robust adjustments for the batch effect on each peptide. In SpectroPipeR a parametric ComBAT emperical Bayes adjustment is implemented by utilizing the sva-package.

covariate adjustment

If a covariate adjustment of peptide intensity data was performed using the users input formula, a linear mixed model (LMM) was calculated based on that formula per peptide and the outcoming residuals were added to the mean peptide intensity over the samples. This means that the adjusted peptide intensities retain their intensity level (low intense peptides keep their low intensity and high intense ions keep their higher intensity).

Examples

# load library
library(SpectroPipeR)

# parameter list
params <- list(output_folder = "../SpectroPipeR_test_folder",
              ion_q_value_cutoff = 0.01,
              id_drop_cutoff = 0.3,
              normalization_method = "median",
              normalization_factor_cutoff_outlier = 4,
              filter_oxidized_peptides = TRUE,
              protein_intensity_estimation = "MaxLFQ",
              stat_test = "modt",
              type_slr = "median",
              fold_change = 1.5,
              p_value_cutoff = 0.05,
              paired = FALSE
             )

# example input file
example_file_path <- system.file("extdata",
                                "SN_test_HYE_mix_file.tsv",
                                package="SpectroPipeR")
# perform the analysis
SpectroPipeR_analysis <- SpectroPipeR(file = example_file_path,
                                     parameter = params,
                                     condition_comparisons = cbind(c("HYE mix A","HYE mix B"))
                                     )
#> #*****************************************
#> # READ SPECTRONAUT MODULE
#> #*****************************************
#> 
#> checking parameters ...
#> loading data ...
#> write input data to output folder ...
#> R.FileName capping ...
#> 02_ID_rate/8_sample_analysis ATTENTION !!! ---> folder already exists - files will be replaced !!!
#> generating intermediate SDRF file
#> saving intermediate SDRF file ... please use https://lessdrf.streamlit.app for finalizing/editing file before submitting to a public repository
#> _________ data set loaded with ... _________
#> number of raw files = 8
#> number of conditions = 2
#> number of ions without filtering = 1095
#> number of peptides without filtering = 904
#> number of Protein groups without filtering = 93
#> count profiled values ...
#> performing counting of min. sample percentage where an ion was detected...
#> ion Q-value cutoff < 0.01
#> 25 % (2/8) is the min. sample percentage where an ion was detected
#> performing ID rate filtering ...
#> performing protein count over replicates (more than or equal 2 peptides) ...
#> performing protein count (<2 and more than or equal 2 peptides) ...
#> _________ ON/OFF analysis: _________
#> ... filter with Q-value 0.01 ...
#> ... filter for 2 peptides and min. present in 50% of replicates ...
#> ... ON/OFF analysis write outputs ...
#> count missed cleavages ...
#> save missed cleavage plots...
#> _________ ID rate per sample: _________
#> ions: median = 1053.5; min. = 1019; max. = 1067
#> modified peptides: median = 921; min. = 893; max. = 931
#> stripped peptides: median = 872; min. = 847; max. = 885
#> protein groups: median = 92.5; min. = 90; max. = 93
#> read module done --> please check outputs in folder: ../SpectroPipeR_test_folder/02_ID_rate/
#> 
#> #*****************************************
#> # NORMALIZATION & QUANTIFICATION MODULE
#> #*****************************************
#> 
#> sorting Replicates and conditions ...
#> NORMALIZATION WAS DONE IN SPECTRONAUT...
#> ...skipping normalization step and use Spectronaut normalized data instead...
#> ../SpectroPipeR_test_folder/03_normalization/8_sample_analysis ATTENTION !!! ---> folder already exists - files will be replaced !!!
#> ../SpectroPipeR_test_folder/05_processed_data/8_sample_analysis ATTENTION !!! ---> folder already exists - files will be replaced !!!
#> save Normalization factor plot ...
#> save Normalization boxplot ...
#> generate ion CV data...
#> save ion CV data plots...
#> save ion CV data vs. mean intensity hexbin plots...
#> calculating peptide intensity data ...
#> writing peptide intensity data ...
#> filter Met_oxidized peptides ... TRUE
#> 60 Met_oxidized peptides filtered for quantification
#> replacing 0.0139821029082774 % of values (ZERO quantity values)
#> with 0.5 = half-minimal value
#> protein intensity calculation ...
#> extracting iBAQ intensities from Spectronaut report ...
#> calc. mean, SD, CV of iBAQ intensities ...
#> ... save iBAQ data ...
#> perform maxLFQ protein intensity calculation ... (this will take some time)
#> ... preprocessing data for MaxLFQ estimation ...
#> Concatenating secondary ids...
#> Removing low intensities...
#> ... generate protein list for MaxLFQ estimation ...
#> # proteins = 93, # samples = 8
#> 5.4%
#> 11%
#> 16%
#> 22%
#> 27%
#> 32%
#> 38%
#> 43%
#> 48%
#> 54%
#> 59%
#> 65%
#> 70%
#> 75%
#> 81%
#> 86%
#> 91%
#> 97%
#> Completed.
#> ... calculation of MaxLFQ ...
#> 5.4%
#> 11%
#> 16%
#> 22%
#> 27%
#> 32%
#> 38%
#> 43%
#> 48%
#> 54%
#> 59%
#> 65%
#> 70%
#> 75%
#> 81%
#> 86%
#> 91%
#> 97%
#> Completed.
#> ... generate outputs for MaxLFQ estimation ...
#> ... do median normalization of MaxLFQ data ...
#> ... save MaxLFQ boxplot ...
#> ... save MaxLFQ data ...
#> ... compare protein intensities and iBAQ protein intensities ...
#> ... CV plot calculation ...
#> ... calc. cumulative frequency of CV ...
#>  [=                                                           ]    0.99% - calculation ...  [=                                                           ]    1.98% - calculation ...  [==                                                          ]    2.97% - calculation ...  [==                                                          ]    3.96% - calculation ...  [===                                                         ]    4.95% - calculation ...  [====                                                        ]    5.94% - calculation ...  [====                                                        ]    6.93% - calculation ...  [=====                                                       ]    7.92% - calculation ...  [=====                                                       ]    8.91% - calculation ...  [======                                                      ]    9.90% - calculation ...  [=======                                                     ]   10.89% - calculation ...  [=======                                                     ]   11.88% - calculation ...  [========                                                    ]   12.87% - calculation ...  [========                                                    ]   13.86% - calculation ...  [=========                                                   ]   14.85% - calculation ...  [==========                                                  ]   15.84% - calculation ...  [==========                                                  ]   16.83% - calculation ...  [===========                                                 ]   17.82% - calculation ...  [===========                                                 ]   18.81% - calculation ...  [============                                                ]   19.80% - calculation ...  [============                                                ]   20.79% - calculation ...  [=============                                               ]   21.78% - calculation ...  [==============                                              ]   22.77% - calculation ...  [==============                                              ]   23.76% - calculation ...  [===============                                             ]   24.75% - calculation ...  [===============                                             ]   25.74% - calculation ...  [================                                            ]   26.73% - calculation ...  [=================                                           ]   27.72% - calculation ...  [=================                                           ]   28.71% - calculation ...  [==================                                          ]   29.70% - calculation ...  [==================                                          ]   30.69% - calculation ...  [===================                                         ]   31.68% - calculation ...  [====================                                        ]   32.67% - calculation ...  [====================                                        ]   33.66% - calculation ...  [=====================                                       ]   34.65% - calculation ...  [=====================                                       ]   35.64% - calculation ...  [======================                                      ]   36.63% - calculation ...  [=======================                                     ]   37.62% - calculation ...  [=======================                                     ]   38.61% - calculation ...  [========================                                    ]   39.60% - calculation ...  [========================                                    ]   40.59% - calculation ...  [=========================                                   ]   41.58% - calculation ...  [==========================                                  ]   42.57% - calculation ...  [==========================                                  ]   43.56% - calculation ...  [===========================                                 ]   44.55% - calculation ...  [===========================                                 ]   45.54% - calculation ...  [============================                                ]   46.53% - calculation ...  [=============================                               ]   47.52% - calculation ...  [=============================                               ]   48.51% - calculation ...  [==============================                              ]   49.50% - calculation ...  [==============================                              ]   50.50% - calculation ...  [===============================                             ]   51.49% - calculation ...  [===============================                             ]   52.48% - calculation ...  [================================                            ]   53.47% - calculation ...  [=================================                           ]   54.46% - calculation ...  [=================================                           ]   55.45% - calculation ...  [==================================                          ]   56.44% - calculation ...  [==================================                          ]   57.43% - calculation ...  [===================================                         ]   58.42% - calculation ...  [====================================                        ]   59.41% - calculation ...  [====================================                        ]   60.40% - calculation ...  [=====================================                       ]   61.39% - calculation ...  [=====================================                       ]   62.38% - calculation ...  [======================================                      ]   63.37% - calculation ...  [=======================================                     ]   64.36% - calculation ...  [=======================================                     ]   65.35% - calculation ...  [========================================                    ]   66.34% - calculation ...  [========================================                    ]   67.33% - calculation ...  [=========================================                   ]   68.32% - calculation ...  [==========================================                  ]   69.31% - calculation ...  [==========================================                  ]   70.30% - calculation ...  [===========================================                 ]   71.29% - calculation ...  [===========================================                 ]   72.28% - calculation ...  [============================================                ]   73.27% - calculation ...  [=============================================               ]   74.26% - calculation ...  [=============================================               ]   75.25% - calculation ...  [==============================================              ]   76.24% - calculation ...  [==============================================              ]   77.23% - calculation ...  [===============================================             ]   78.22% - calculation ...  [================================================            ]   79.21% - calculation ...  [================================================            ]   80.20% - calculation ...  [=================================================           ]   81.19% - calculation ...  [=================================================           ]   82.18% - calculation ...  [==================================================          ]   83.17% - calculation ...  [==================================================          ]   84.16% - calculation ...  [===================================================         ]   85.15% - calculation ...  [====================================================        ]   86.14% - calculation ...  [====================================================        ]   87.13% - calculation ...  [=====================================================       ]   88.12% - calculation ...  [=====================================================       ]   89.11% - calculation ...  [======================================================      ]   90.10% - calculation ...  [=======================================================     ]   91.09% - calculation ...  [=======================================================     ]   92.08% - calculation ...  [========================================================    ]   93.07% - calculation ...  [========================================================    ]   94.06% - calculation ...  [=========================================================   ]   95.05% - calculation ...  [==========================================================  ]   96.04% - calculation ...  [==========================================================  ]   97.03% - calculation ...  [=========================================================== ]   98.02% - calculation ...  [=========================================================== ]   99.01% - calculation ...  [============================================================]  100.00% - calculation ... 
#> ... render CV plot ...
#> generating raw file / condition file ...
#> no outlier detected with 4 fold difference from the median
#> norm. & quant. module done --> please check outputs in folder: ../SpectroPipeR_test_folder/03_normalization/
#> norm. & quant. module done --> please check outputs in folder: ../SpectroPipeR_test_folder/05_processed_data/
#> 
#> #*****************************************
#> # MVA MODULE
#> #*****************************************
#> 
#> perfoming multivariate analysis ...
#> ../SpectroPipeR_test_folder/04_multivariate_analysis/8_sample_analysis ATTENTION !!! ---> folder already exists - files will be replaced !!!
#> generating raw file / condition file ...
#> PCA analysis on peptide level ...
#> PCA analysis on protein level ...
#> ... 0 proteins removed due to NAs
#> generating Scree plots ...
#> generating Top10 variable contribution plot ...
#> generating 3D PCA plot ...
#> generating PCA plot 1st vs. 2nd dimension ...
#> generating PCA plot 1st to 5th dimension ...
#> saving PCA plot 1st to 5th dimension ...
#> generating Biplot of Top10 variables and individuals ...
#> generating individuals contribution plot ...
#> generating PCA plot matrix ...
#> saving PCA plot matrix ...
#> saving 3D PCA plots ...
#> performing correlation analysis ...
#> performing correlation analysis - calculate p-values ...
#> generating correlation plots ...
#> saving correlation plots ...
#> saving correlation plots with condition info ...
#> perform UMAP analysis on protein level ...
#> try UMAP with default settings ...
#> due to low complexity try UMAP with n_neighbors = 5 setting ...
#> MVA module done --> please check outputs in folder: ../SpectroPipeR_test_folder/04_multivariate_analysis/
#> 
#> #*****************************************
#> # STATISTICS MODULE
#> #*****************************************
#> 
#> ../SpectroPipeR_test_folder/06_statistics/8_sample_analysis ATTENTION !!! ---> folder already exists - files will be replaced !!!
#> reformatting data ...
#> register processor cores ...
#> performing statistical analysis (this might take a while) ...
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> close cores ...
#> 
#> start to end time comparison for stat. analysis: 0.00158582303259108 hours
#> estimating effect sizes ...
#>  [============================================================]  100.00% - calc. effect sizes... 
#> 
#> join and tidy tables ...
#> filtering statistical table using supplied cutoffs ...
#> writing output files ...
#> adding iBAQ quantiles to statistics table ...
#> generating Excel outputs ...
#> performing fold-change cutoff sensitivity analysis ...
#> plotting fold-change cutoff sensitivity analysis ...
#>  [================================================================================]  100.00% - 1 
#> 
#> plotting fold-change simple cutoff sensitivity analysis (peptide n > 1)...
#>  [================================================================================]  100.00% - 1 
#> 
#> generating volcano plots ...
#>  [============================================================]  100.00% - Volcano plots with adj. p-value 
#>  [============================================================]  100.00% - Volcano plots with adj. p-value 
#>  [============================================================]  100.00% - Volcano plots with raw p-value 
#> 
#> condition-comparison-wise signal to noise comparison ...
#> ...signal to noise: save scatter plot...
#> condition-comparison-wise comparison of peptide-int.-ratios vs. protein-int.-ratios ...
#> ...calculating protein ratios...
#> ...combining ratio tables...
#> ...generating ratio-ratios: protein_ratios/peptide_ratios...
#> ...filter for at least 2 peptides...
#> ...adding of protein intensity to table...
#> ...add signal to noise per group...
#> ...add detection with selected q-value cutoff with at least 2 peptides per replicate...
#> ...add direction comparison for protein or peptide condition comp. ratio...
#> ...save table of stat. significant with poor signal to noise...
#> ...counting protein which having a 2fold difference...
#> ...select Top15 over- or under-estimated proteins...
#> ...protein int. benchmark: save scatter plot...
#> ...protein int. benchmark: save histogram plot...
#> ...protein int. benchmark: save table...
#> ...counting protein: gradient of difference...
#> ...protein int. benchmark: gradient of difference area plots...
#> statistics module done --> please check outputs in folder: ../SpectroPipeR_test_folder/06_statistics/
#> 
#> #*****************************************
#> # REPORT MODULE
#> #*****************************************
#> 
#> generating methods part ...
#> render HTML report ... this might take a while
#> 
#> 
#> processing file: DIA_MS_analysis_report_Master.qmd
#> 1/37                   
#> 2/37 [global.options]  
#> 3/37                   
#> 4/37 [setup]           
#> 5/37                   
#> 6/37 [unnamed-chunk-1] 
#> 7/37                   
#> 8/37 [unnamed-chunk-2] 
#> 9/37                   
#> 10/37 [unnamed-chunk-3] 
#> 11/37                   
#> # A tibble: 1 × 1
#>   value
#>   <chr>
#> 1 0.01 
#> 12/37 [unnamed-chunk-4] 
#> 13/37                   
#> 14/37 [unnamed-chunk-5] 
#> 15/37                   
#> 16/37 [unnamed-chunk-6] 
#> 17/37                   
#> 18/37 [unnamed-chunk-7] 
#> 19/37                   
#> 20/37 [unnamed-chunk-8] 
#> 21/37                   
#> 22/37 [unnamed-chunk-9] 
#> 23/37                   
#> 24/37 [unnamed-chunk-10]
#> 25/37                   
#> 26/37 [unnamed-chunk-11]
#> 27/37                   
#> 28/37 [unnamed-chunk-12]
#> 29/37                   
#> 30/37 [unnamed-chunk-13]
#> 31/37                   
#> 32/37 [unnamed-chunk-14]
#> 33/37                   
#> 34/37 [unnamed-chunk-15]
#> 35/37                   
#> 36/37 [unnamed-chunk-16]
#> 37/37                   
#> output file: DIA_MS_analysis_report_Master.knit.md
#> 
#> pandoc --output 2024_06_28__15_48_SpectroPipeR_report.html
#>   to: html
#>   standalone: true
#>   self-contained: true
#>   section-divs: true
#>   html-math-method: katex
#>   wrap: none
#>   default-image-extension: png
#>   css:
#>     - styles.css
#>   toc: true
#>   toc-depth: 3
#>   
#> metadata
#>   document-css: false
#>   link-citations: true
#>   date-format: long
#>   lang: en
#>   title: SpectroPipeR DIA-MS analysis report
#>   author: Stephan Michalik
#>   date: '`r format(Sys.Date(), "%B %d, %Y")`'
#>   title-block-banner: '#151515'
#>   subtitle: automatically generated report
#>   page-layout: full
#>   toc-title: Contents
#>   theme: united
#>   highlight: tango
#>   df_print: paged
#>   toc-location: left
#>   anchor-sections: true
#>   smooth-scroll: true
#>   
#> Output created: 2024_06_28__15_48_SpectroPipeR_report.html
#> 
#> 
#> render HTML report ... DONE!