SpectroPipeR: report module

Function for generating an interactive analysis report (HTML file). This is the final part of the SpectroPipeR pipeline where the analysis modules output serves as input.

Usage

SpectroPipeR_report_module(
  SpectroPipeR_data = NULL,
  SpectroPipeR_data_quant = NULL,
  SpectroPipeR_data_stats = NULL,
  open_rendered_report = FALSE
)

Arguments

SpectroPipeR_data

(mandatory) is the SpectroPipeR_data list object from read_spectronaut_module() object e.g. SpectroPipeR_data please see example below

SpectroPipeR_data_quant

(mandatory) is the SpectroPipeR_data_quant list object from norm_quant_module() object e.g. SpectroPipeR_data_quant please see example below

SpectroPipeR_data_stats

(optional) is the SpectroPipeR_data_quant list object from statistics_module() object e.g. SpectroPipeR_data_stats please see example below; if you have only 1 replicate data set a statistical analysis is not possible so leave this parameter to NULL a report without statitsics will be generated

open_rendered_report

logical - if rendered report should be opened (TRUE or FALSE); default = FALSE

Value

generates and exports an interactive standalone html report to the output folder

Examples

#load library
library(SpectroPipeR)

# parameter list
params <- list(output_folder = "../SpectroPipeR_test_folder",
              ion_q_value_cutoff = 0.01,
              id_drop_cutoff = 0.3,
              normalization_method = "median",
              normalization_factor_cutoff_outlier = 4,
              filter_oxidized_peptides = TRUE,
              protein_intensity_estimation = "MaxLFQ",
              stat_test = "modt",
              type_slr = "median",
              fold_change = 1.5,
              p_value_cutoff = 0.05,
              paired = FALSE
)

# example input file
example_file_path <- system.file("extdata",
                                "SN_test_HYE_mix_file.tsv",
                                 package="SpectroPipeR")

# step 1: load Spectronaut data module
SpectroPipeR_data <- read_spectronaut_module(file = example_file_path,
                                            parameter = params,
                                            print.plot = FALSE)
#> #*****************************************
#> # READ SPECTRONAUT MODULE
#> #*****************************************
#> 
#> checking parameters ...
#> loading data ...
#> write input data to output folder ...
#> R.FileName capping ...
#> 02_ID_rate/8_sample_analysis ATTENTION !!! ---> folder already exists - files will be replaced !!!
#> generating intermediate SDRF file
#> saving intermediate SDRF file ... please use https://lessdrf.streamlit.app for finalizing/editing file before submitting to a public repository
#> _________ data set loaded with ... _________
#> number of raw files = 8
#> number of conditions = 2
#> number of ions without filtering = 1095
#> number of peptides without filtering = 904
#> number of Protein groups without filtering = 93
#> count profiled values ...
#> performing counting of min. sample percentage where an ion was detected...
#> ion Q-value cutoff < 0.01
#> 25 % (2/8) is the min. sample percentage where an ion was detected
#> performing ID rate filtering ...
#> performing protein count over replicates (more than or equal 2 peptides) ...
#> performing protein count (<2 and more than or equal 2 peptides) ...
#> _________ ON/OFF analysis: _________
#> ... filter with Q-value 0.01 ...
#> ... filter for 2 peptides and min. present in 50% of replicates ...
#> ... ON/OFF analysis write outputs ...
#> count missed cleavages ...
#> save missed cleavage plots...
#> _________ ID rate per sample: _________
#> ions: median = 1053.5; min. = 1019; max. = 1067
#> modified peptides: median = 921; min. = 893; max. = 931
#> stripped peptides: median = 872; min. = 847; max. = 885
#> protein groups: median = 92.5; min. = 90; max. = 93
#> read module done --> please check outputs in folder: ../SpectroPipeR_test_folder/02_ID_rate/

# step 2: normalize & quantification module
SpectroPipeR_data_quant <- norm_quant_module(SpectroPipeR_data = SpectroPipeR_data)
#> 
#> #*****************************************
#> # NORMALIZATION & QUANTIFICATION MODULE
#> #*****************************************
#> 
#> sorting Replicates and conditions ...
#> NORMALIZATION WAS DONE IN SPECTRONAUT...
#> ...skipping normalization step and use Spectronaut normalized data instead...
#> ../SpectroPipeR_test_folder/03_normalization/8_sample_analysis ATTENTION !!! ---> folder already exists - files will be replaced !!!
#> ../SpectroPipeR_test_folder/05_processed_data/8_sample_analysis ATTENTION !!! ---> folder already exists - files will be replaced !!!
#> save Normalization factor plot ...
#> save Normalization boxplot ...
#> generate ion CV data...
#> save ion CV data plots...
#> save ion CV data vs. mean intensity hexbin plots...
#> calculating peptide intensity data ...
#> writing peptide intensity data ...
#> filter Met_oxidized peptides ... TRUE
#> 60 Met_oxidized peptides filtered for quantification
#> replacing 0.0139821029082774 % of values (ZERO quantity values)
#> with 0.5 = half-minimal value
#> protein intensity calculation ...
#> extracting iBAQ intensities from Spectronaut report ...
#> calc. mean, SD, CV of iBAQ intensities ...
#> ... save iBAQ data ...
#> perform maxLFQ protein intensity calculation ... (this will take some time)
#> ... preprocessing data for MaxLFQ estimation ...
#> Concatenating secondary ids...
#> Removing low intensities...
#> ... generate protein list for MaxLFQ estimation ...
#> # proteins = 93, # samples = 8
#> 5.4%
#> 11%
#> 16%
#> 22%
#> 27%
#> 32%
#> 38%
#> 43%
#> 48%
#> 54%
#> 59%
#> 65%
#> 70%
#> 75%
#> 81%
#> 86%
#> 91%
#> 97%
#> Completed.
#> ... calculation of MaxLFQ ...
#> 5.4%
#> 11%
#> 16%
#> 22%
#> 27%
#> 32%
#> 38%
#> 43%
#> 48%
#> 54%
#> 59%
#> 65%
#> 70%
#> 75%
#> 81%
#> 86%
#> 91%
#> 97%
#> Completed.
#> ... generate outputs for MaxLFQ estimation ...
#> ... do median normalization of MaxLFQ data ...
#> ... save MaxLFQ boxplot ...
#> ... save MaxLFQ data ...
#> ... compare protein intensities and iBAQ protein intensities ...
#> ... CV plot calculation ...
#> ... calc. cumulative frequency of CV ...
#>  [=                                                           ]    0.99% - calculation ...  [=                                                           ]    1.98% - calculation ...  [==                                                          ]    2.97% - calculation ...  [==                                                          ]    3.96% - calculation ...  [===                                                         ]    4.95% - calculation ...  [====                                                        ]    5.94% - calculation ...  [====                                                        ]    6.93% - calculation ...  [=====                                                       ]    7.92% - calculation ...  [=====                                                       ]    8.91% - calculation ...  [======                                                      ]    9.90% - calculation ...  [=======                                                     ]   10.89% - calculation ...  [=======                                                     ]   11.88% - calculation ...  [========                                                    ]   12.87% - calculation ...  [========                                                    ]   13.86% - calculation ...  [=========                                                   ]   14.85% - calculation ...  [==========                                                  ]   15.84% - calculation ...  [==========                                                  ]   16.83% - calculation ...  [===========                                                 ]   17.82% - calculation ...  [===========                                                 ]   18.81% - calculation ...  [============                                                ]   19.80% - calculation ...  [============                                                ]   20.79% - calculation ...  [=============                                               ]   21.78% - calculation ...  [==============                                              ]   22.77% - calculation ...  [==============                                              ]   23.76% - calculation ...  [===============                                             ]   24.75% - calculation ...  [===============                                             ]   25.74% - calculation ...  [================                                            ]   26.73% - calculation ...  [=================                                           ]   27.72% - calculation ...  [=================                                           ]   28.71% - calculation ...  [==================                                          ]   29.70% - calculation ...  [==================                                          ]   30.69% - calculation ...  [===================                                         ]   31.68% - calculation ...  [====================                                        ]   32.67% - calculation ...  [====================                                        ]   33.66% - calculation ...  [=====================                                       ]   34.65% - calculation ...  [=====================                                       ]   35.64% - calculation ...  [======================                                      ]   36.63% - calculation ...  [=======================                                     ]   37.62% - calculation ...  [=======================                                     ]   38.61% - calculation ...  [========================                                    ]   39.60% - calculation ...  [========================                                    ]   40.59% - calculation ...  [=========================                                   ]   41.58% - calculation ...  [==========================                                  ]   42.57% - calculation ...  [==========================                                  ]   43.56% - calculation ...  [===========================                                 ]   44.55% - calculation ...  [===========================                                 ]   45.54% - calculation ...  [============================                                ]   46.53% - calculation ...  [=============================                               ]   47.52% - calculation ...  [=============================                               ]   48.51% - calculation ...  [==============================                              ]   49.50% - calculation ...  [==============================                              ]   50.50% - calculation ...  [===============================                             ]   51.49% - calculation ...  [===============================                             ]   52.48% - calculation ...  [================================                            ]   53.47% - calculation ...  [=================================                           ]   54.46% - calculation ...  [=================================                           ]   55.45% - calculation ...  [==================================                          ]   56.44% - calculation ...  [==================================                          ]   57.43% - calculation ...  [===================================                         ]   58.42% - calculation ...  [====================================                        ]   59.41% - calculation ...  [====================================                        ]   60.40% - calculation ...  [=====================================                       ]   61.39% - calculation ...  [=====================================                       ]   62.38% - calculation ...  [======================================                      ]   63.37% - calculation ...  [=======================================                     ]   64.36% - calculation ...  [=======================================                     ]   65.35% - calculation ...  [========================================                    ]   66.34% - calculation ...  [========================================                    ]   67.33% - calculation ...  [=========================================                   ]   68.32% - calculation ...  [==========================================                  ]   69.31% - calculation ...  [==========================================                  ]   70.30% - calculation ...  [===========================================                 ]   71.29% - calculation ...  [===========================================                 ]   72.28% - calculation ...  [============================================                ]   73.27% - calculation ...  [=============================================               ]   74.26% - calculation ...  [=============================================               ]   75.25% - calculation ...  [==============================================              ]   76.24% - calculation ...  [==============================================              ]   77.23% - calculation ...  [===============================================             ]   78.22% - calculation ...  [================================================            ]   79.21% - calculation ...  [================================================            ]   80.20% - calculation ...  [=================================================           ]   81.19% - calculation ...  [=================================================           ]   82.18% - calculation ...  [==================================================          ]   83.17% - calculation ...  [==================================================          ]   84.16% - calculation ...  [===================================================         ]   85.15% - calculation ...  [====================================================        ]   86.14% - calculation ...  [====================================================        ]   87.13% - calculation ...  [=====================================================       ]   88.12% - calculation ...  [=====================================================       ]   89.11% - calculation ...  [======================================================      ]   90.10% - calculation ...  [=======================================================     ]   91.09% - calculation ...  [=======================================================     ]   92.08% - calculation ...  [========================================================    ]   93.07% - calculation ...  [========================================================    ]   94.06% - calculation ...  [=========================================================   ]   95.05% - calculation ...  [==========================================================  ]   96.04% - calculation ...  [==========================================================  ]   97.03% - calculation ...  [=========================================================== ]   98.02% - calculation ...  [=========================================================== ]   99.01% - calculation ...  [============================================================]  100.00% - calculation ... 
#> ... render CV plot ...
#> generating raw file / condition file ...
#> no outlier detected with 4 fold difference from the median
#> norm. & quant. module done --> please check outputs in folder: ../SpectroPipeR_test_folder/03_normalization/
#> norm. & quant. module done --> please check outputs in folder: ../SpectroPipeR_test_folder/05_processed_data/

# step 3: MVA module
SpectroPipeR_MVA <- MVA_module(SpectroPipeR_data_quant = SpectroPipeR_data_quant,
          HCPC_analysis = FALSE)
#> 
#> #*****************************************
#> # MVA MODULE
#> #*****************************************
#> 
#> perfoming multivariate analysis ...
#> ../SpectroPipeR_test_folder/04_multivariate_analysis/8_sample_analysis ATTENTION !!! ---> folder already exists - files will be replaced !!!
#> generating raw file / condition file ...
#> PCA analysis on peptide level ...
#> PCA analysis on protein level ...
#> ... 0 proteins removed due to NAs
#> generating Scree plots ...
#> generating Top10 variable contribution plot ...
#> generating 3D PCA plot ...
#> generating PCA plot 1st vs. 2nd dimension ...
#> generating PCA plot 1st to 5th dimension ...
#> saving PCA plot 1st to 5th dimension ...
#> generating Biplot of Top10 variables and individuals ...
#> generating individuals contribution plot ...
#> generating PCA plot matrix ...
#> saving PCA plot matrix ...
#> saving 3D PCA plots ...
#> performing correlation analysis ...
#> performing correlation analysis - calculate p-values ...
#> generating correlation plots ...
#> saving correlation plots ...
#> saving correlation plots with condition info ...
#> perform UMAP analysis on protein level ...
#> try UMAP with default settings ...
#> due to low complexity try UMAP with n_neighbors = 5 setting ...
#> MVA module done --> please check outputs in folder: ../SpectroPipeR_test_folder/04_multivariate_analysis/

# step 4: statistics module
SpectroPipeR_data_stats <- statistics_module(SpectroPipeR_data_quant = SpectroPipeR_data_quant,
                                            condition_comparisons = cbind(c("HYE mix A",
                                                                            "HYE mix B")))
#> 
#> #*****************************************
#> # STATISTICS MODULE
#> #*****************************************
#> 
#> ../SpectroPipeR_test_folder/06_statistics/8_sample_analysis ATTENTION !!! ---> folder already exists - files will be replaced !!!
#> reformatting data ...
#> register processor cores ...
#> performing statistical analysis (this might take a while) ...
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> 
#> close cores ...
#> 
#> start to end time comparison for stat. analysis: 0.00167782141102685 hours
#> estimating effect sizes ...
#>  [============================================================]  100.00% - calc. effect sizes... 
#> 
#> join and tidy tables ...
#> filtering statistical table using supplied cutoffs ...
#> writing output files ...
#> adding iBAQ quantiles to statistics table ...
#> generating Excel outputs ...
#> performing fold-change cutoff sensitivity analysis ...
#> plotting fold-change cutoff sensitivity analysis ...
#>  [================================================================================]  100.00% - 1 
#> 
#> plotting fold-change simple cutoff sensitivity analysis (peptide n > 1)...
#>  [================================================================================]  100.00% - 1 
#> 
#> generating volcano plots ...
#>  [============================================================]  100.00% - Volcano plots with adj. p-value 
#>  [============================================================]  100.00% - Volcano plots with adj. p-value 
#>  [============================================================]  100.00% - Volcano plots with raw p-value 
#> 
#> condition-comparison-wise signal to noise comparison ...
#> ...signal to noise: save scatter plot...
#> condition-comparison-wise comparison of peptide-int.-ratios vs. protein-int.-ratios ...
#> ...calculating protein ratios...
#> ...combining ratio tables...
#> ...generating ratio-ratios: protein_ratios/peptide_ratios...
#> ...filter for at least 2 peptides...
#> ...adding of protein intensity to table...
#> ...add signal to noise per group...
#> ...add detection with selected q-value cutoff with at least 2 peptides per replicate...
#> ...add direction comparison for protein or peptide condition comp. ratio...
#> ...save table of stat. significant with poor signal to noise...
#> ...counting protein which having a 2fold difference...
#> ...select Top15 over- or under-estimated proteins...
#> ...protein int. benchmark: save scatter plot...
#> ...protein int. benchmark: save histogram plot...
#> ...protein int. benchmark: save table...
#> ...counting protein: gradient of difference...
#> ...protein int. benchmark: gradient of difference area plots...
#> statistics module done --> please check outputs in folder: ../SpectroPipeR_test_folder/06_statistics/

# step 5: report module
SpectroPipeR_report_module(SpectroPipeR_data = SpectroPipeR_data,
                          SpectroPipeR_data_quant = SpectroPipeR_data_quant,
                          SpectroPipeR_data_stats = SpectroPipeR_data_stats)
#> 
#> #*****************************************
#> # REPORT MODULE
#> #*****************************************
#> 
#> generating methods part ...
#> render HTML report ... this might take a while
#> 
#> 
#> processing file: DIA_MS_analysis_report_Master.qmd
#> 1/37                   
#> 2/37 [global.options]  
#> 3/37                   
#> 4/37 [setup]           
#> 5/37                   
#> 6/37 [unnamed-chunk-1] 
#> 7/37                   
#> 8/37 [unnamed-chunk-2] 
#> 9/37                   
#> 10/37 [unnamed-chunk-3] 
#> 11/37                   
#> # A tibble: 1 × 1
#>   value
#>   <chr>
#> 1 0.01 
#> 12/37 [unnamed-chunk-4] 
#> 13/37                   
#> 14/37 [unnamed-chunk-5] 
#> 15/37                   
#> 16/37 [unnamed-chunk-6] 
#> 17/37                   
#> 18/37 [unnamed-chunk-7] 
#> 19/37                   
#> 20/37 [unnamed-chunk-8] 
#> 21/37                   
#> 22/37 [unnamed-chunk-9] 
#> 23/37                   
#> 24/37 [unnamed-chunk-10]
#> 25/37                   
#> 26/37 [unnamed-chunk-11]
#> 27/37                   
#> 28/37 [unnamed-chunk-12]
#> 29/37                   
#> 30/37 [unnamed-chunk-13]
#> 31/37                   
#> 32/37 [unnamed-chunk-14]
#> 33/37                   
#> 34/37 [unnamed-chunk-15]
#> 35/37                   
#> 36/37 [unnamed-chunk-16]
#> 37/37                   
#> output file: DIA_MS_analysis_report_Master.knit.md
#> 
#> pandoc --output 2024_06_28__15_50_SpectroPipeR_report.html
#>   to: html
#>   standalone: true
#>   self-contained: true
#>   section-divs: true
#>   html-math-method: katex
#>   wrap: none
#>   default-image-extension: png
#>   css:
#>     - styles.css
#>   toc: true
#>   toc-depth: 3
#>   
#> metadata
#>   document-css: false
#>   link-citations: true
#>   date-format: long
#>   lang: en
#>   title: SpectroPipeR DIA-MS analysis report
#>   author: Stephan Michalik
#>   date: '`r format(Sys.Date(), "%B %d, %Y")`'
#>   title-block-banner: '#151515'
#>   subtitle: automatically generated report
#>   page-layout: full
#>   toc-title: Contents
#>   theme: united
#>   highlight: tango
#>   df_print: paged
#>   toc-location: left
#>   anchor-sections: true
#>   smooth-scroll: true
#>   
#> Output created: 2024_06_28__15_50_SpectroPipeR_report.html
#> 
#> 
#> render HTML report ... DONE!