Extraction and plotting of XICs from Spectronaut's exported SQLite Databases (Spectronaut version > 18.7.240506.55695; XIC DB export generates one SQLite db file per raw file)

Exports user-selected protein group XIC (MS1/MS2) data in both PDF and CSV formats for all detected ions.

Usage

XIC_plot_module(
  Spectronaut_report_path = NULL,
  Spectronaut_xicDB_path = NULL,
  protein_groups = NULL,
  output_path = NULL,
  export_csv_files = FALSE,
  run_specific_y_axis = FALSE,
  selected_conditions = NULL,
  number_of_cores = 2
)

Arguments

Spectronaut_report_path

file; Path pointing to Spectronaut analysis report using SpectroPipeR report export scheme.

Spectronaut_xicDB_path

folder; Path pointing to Spectronaut all XIC SQLite database export folder from that analysis (use pipeline mode in Spectronaut).

protein_groups

Vector containing user-selected protein group IDs for generating plots. e.g. c("P38720","P29311")

output_path

Define the output directory for the generated ion-specific plots (PDF) and data tables (CSV) for user-selected protein groups.

export_csv_files

Export individual CSV files per ion (boolean, default: FALSE).

run_specific_y_axis

if the XIC plot should be globally scaled y-axis or run specific (default: FALSE)

selected_conditions

Vector of user-selected conditions for plotting (serves for both selection and ordering).

number_of_cores

Define the number of CPU cores to utilize for parallel XIC extraction and plotting.

Value

Exports user-selected protein group XIC data (MS1/MS2) in both PDF and CSV formats for all detected ions of a protein.

Details

Description of the Spectronaut scores shown in the output plots:

parameter 	 description

EG.Qvalue	The Q-value is the multiple test corrected p-value. The Q-value can be used

as an estimate for the FDR at a given Cscore cutoff.

EG.PEP	The posterior error probability for a peptide precursor observation. The PEP

is calculated as the decoy probability function for a specficic Cscore divided

by the probability function for decoy plus target. PEP = d/(t+d) where d is the

value of the decoy probability function for a given Cscore and t is the value for

target probability function for a given Cscore.

FG.ShapeQualityScore	The average of all available peak quality measurements

(scale: 1.0 best to -1.0 worst)

EG.IntCorrScore	The correlation score between the relative fragment intensities as provided

by the spectral librarxy and the empirical observed fragtment intensitiers

at XIC peak apex.

EG.SignalToNoise	The signal to noise ratio of the LC-MS peak for a given peptide precursor.

The signal is calc. as the maximum intensity of the fragment sum XIC within

the peak boundaries. The noise is calculated as the average fragment sum XIC

intensity outside the peak boundaries.

EG.Cscore	The discriminant score used for identification. This score is a linear combination

of all applicable scores for a certain workflow.

EG.Identified	Returns weather a precursor is considered identified or not. In order to be

identified it has to match both the precursor and the protein FDR cutoff.

dppp (MS1)	Returns the number of data points (scans) that make up the MS1 XIC of the selected

LC-MS peak for this precursor. Only data points that lie within the start and the

end RT are considered.

dppp (MS2)	Returns the number of data points (scans) that make up the MS2 XIC of the selected

LC-MS peak for this precursor. Only data points that lie within the start and the

end RT are considered.

The vertical dashed line in the plots indicate the individual ApexRT, which is calculated based on all fragment XIC for a specific peptide precursor ion.

Examples

#setup example input paths / protein selection for plotting
Spectronaut_report_path <- system.file("extdata/HYE_demo_data",
                                       "HYE_demo_data_Report_SpectroPipeR.tsv",
                                        package="SpectroPipeR")
Spectronaut_xicDB_path <- system.file("extdata/HYE_demo_data/XIC_DBs",package="SpectroPipeR")
protein_groups <- c("P29311","P38720")
output_path <- "../SpectroPipeR_test_folder/single_XIC_plots"

# extracting and plotting of XIC
XIC_plot_module(Spectronaut_report_path = Spectronaut_report_path,
                Spectronaut_xicDB_path = Spectronaut_xicDB_path,
                protein_groups = protein_groups,
                output_path = output_path,
                number_of_cores = 2
)