SpectroPipeR: statistics module

Function for performing statistical analysis on Spectronaut output reports, representing the fourth step in the pipeline and is dependent on the normalization and quantification module output from step 2.

Usage

statistics_module(
  SpectroPipeR_data_quant = NULL,
  condition_comparisons = NULL,
  number_of_cores = 2
)

Arguments

SpectroPipeR_data_quant

it is the SpectroPipeR_data_quant list object from norm_quant_module() object e.g. SpectroPipeR_data_quant see example below

condition_comparisons

condition comparisons for pairwise- comparison; e.g. condition_comparisons <- cbind(c("condition1","control"),c("condition3","control") )

number_of_cores

number of processor cores to be used for the calculations default = 2; parallel::detectCores()-2 for faster processing (will detect the number of cores in the system and use nearly all cores)

Value

SpectroPipeR_statistics list element containing the statistics analysis results in addition to the automatically generated plots and tables in output folder For the description of the generated figures and tables please read the manual & vignettes

list element 	 description

stat_results	tibble: statistical analysis results table

stat_column_description	tibble: statistical analysis results table column description

stats_results_iBAQ_quantiles	tibble: statistical analysis results table containing

the iBAQ quantilies (Q1-Q10) of the protein per group for a better

ratio judgement

stat_results_filtered	tibble: filtered (user defined FC and adj. p-value) statistical

analysis results table

Examples

#load library
library(SpectroPipeR)

# use default parameters list
params <- list(output_folder = "../SpectroPipeR_test_folder")

# example input file
example_file_path <- system.file("extdata",
                                "SN_test_HYE_mix_file.tsv",
                                package="SpectroPipeR")

# step 1: load Spectronaut data module
SpectroPipeR_data <- read_spectronaut_module(file = example_file_path,
                                            parameter = params,
                                            print.plot = FALSE)

# step 2: normalize & quantification module
SpectroPipeR_data_quant <- norm_quant_module(SpectroPipeR_data = SpectroPipeR_data)

# step 3: MVA module
SpectroPipeR_data_MVA <- MVA_module(SpectroPipeR_data_quant = SpectroPipeR_data_quant,
          HCPC_analysis = FALSE)

# step 4: statistics module
SpectroPipeR_data_stats <- statistics_module(SpectroPipeR_data_quant = SpectroPipeR_data_quant,
                                            condition_comparisons = cbind(c("HYE mix A",
                                                                            "HYE mix B")
                                                                          )
                                            )