SpectroPipeR: loading Spectronaut data module

Function for loading Spectronaut data and performing identification (ID) analysis, which is an essential first step in the SpectroPipeR workflow.

Usage

read_spectronaut_module(
  file = "",
  ID_condition_filtering = FALSE,
  ID_condition_filtering_percent = 0.5,
  parameter = list(),
  max_chars_file_name_capping = 35,
  print.plot = FALSE
)

Arguments

file

location (path) of Spectronaut output report; you should use the Spectronaut_export_scheme() function for getting a SpectroPipeR report scheme encompassing all mandatory columns

ID_condition_filtering

TRUE or FALSE if a condition-wise filtering should be performed

ID_condition_filtering_percent

(numerical value ranging from 0 - 1, default = 0.5) define the proportion for the condition-wise ID filtering

parameter

mandatory parameter list element

table of list elements:

parameter 	 description

output_folder	mandatory !!! - character - output folder path (abs.)

ion_q_value_cutoff	default = 0.01 - numeric - Q-value used in Spectronaut analysis: Biognosys

default is 0.01 = 1% error rate

id_drop_cutoff	default = 0.3 - numeric - value between 0-1 (1 = 100%); xx percent lower

than median of ion ID rate => outlier

normalization_method	default = "median" - character - "median" or Spectronaut - auto-detection

is per default ON, meaning if normalization was performed in Spectronaut

this will be detected and preferred over parameter setting here;

median normalization is the fallback option

normalization_factor_cutoff_outlier	default = 4 - numeric - median off from global median

(4 means abs. 4fold off)

filter_oxidized_peptides	default = TRUE logical - if oxidized peptides should be removed before

peptide quantification

protein_intensity_estimation	default = "MaxLFQ" - character - Hi3 = Hi3 protein intensity estimation,

MaxLFQ = MaxLFQ protein intensity estimation

stat_test	default = "rots" - character - choose statistical test: "rots" = reproducibility

optimized test statistics, "modt" = moderate t-test (lmfit, eBayes),

"t" = t-test

type_slr	default = "median" - character - choose ratio aggregation method:

"median" or "tukey" is used when calculating protein values

fold_change	default = 1.5 - numeric - fold-change used as cutoff e.g. 1.5

p_value_cutoff	default = 0.05 - numeric - p-value used as cutoff e.g. 0.05

paired	default = FALSE - logical - Should paired statistics be applied?

example parameters list (default):

params <- list(output_folder = "../Spectronaut_example",

ion_q_value_cutoff = 0.01,

id_drop_cutoff = 0.3,

normalization_method = "median",

normalization_factor_cutoff_outlier = 4,

filter_oxidized_peptides = T,

protein_intensity_estimation = "MaxLFQ",

stat_test = "rots",

type_slr = "median",

fold_change = 1.5,

p_value_cutoff = 0.05,

paired = FALSE

)

max_chars_file_name_capping

integer, (default = 35) number of max characters used for raw file name presentation; must be adjusted if function

print.plot

if TRUE --> printing ID plot on ion level coloring corresponds ID outlier estimate by "id_drop_cutoff" variable

Value

SpectroPipeR_data list object with the loaded raw data and processed data tables, in addition to the automatically saved tables and plots. For the description of the generated figures and tables please read the manual & vignettes

list element 	 description

spectronaut_output	tibble: Spectronaut report tibble provided for the analysis

SDRF_file	tibble: intermediate SDRF table of the analysis

summary_distinct	tibble: distinct ion, modified peptide, stripped peptides and

protein group count per file filtered by provided Q-value

raw_file_names	tibble: R.FileNames capped and uncapped version together with

R.Condition and R.Replicate

ion_id_median	numerical value: median of ion intensity

ion_id_cutoff	numerical value: ion ID count threshold to classify sample as outlier

PG_2_peptides_ID_raw	tibble: with protein groups with at least 2 peptides with peptide

and replicate count

summary_distinct_outlier	tibble: if outlier are detected they are listed in this tibble

ID_rate_plot	ggplot2 plot: ID rate plot

ID_rate_plot_filter	ggplot2 plot: ion ID rate plot with ion ID cutoff line

sample_length	numberical value: number of samples in the provided Spectronaut report

parameter	list: parameters provided by the user

time_stamp_log_file	string: time stamp of the log file (format: %Y_%m_%d__%H_%M)

log_file_name	string: analysis log file name

Examples

#load library
library(SpectroPipeR)

# use default parameters list
params <- list(output_folder = "../SpectroPipeR_test_folder")

# example input file
example_file_path <- system.file("extdata",
                                "SN_test_HYE_mix_file.tsv",
                                package="SpectroPipeR")

# step 1: load Spectronaut data module
SpectroPipeR_data <- read_spectronaut_module(file = example_file_path,
                                            parameter = params,
                                            print.plot = FALSE)